% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drfDLNMadditive.R
\name{drfDLNMadditive}
\alias{drfDLNMadditive}
\title{Title Fit a single index distributed response function distributed lag non-linear model
(SID-DLNM)}
\usage{
drfDLNMadditive(
  formula,
  smooth = NULL,
  unpen.smooth = NULL,
  fe.varying = NULL,
  fe.cont = NULL,
  offset = NULL,
  dat = NULL,
  maxL = 14,
  kE = 20,
  kw = 10,
  derivH.nthreads = 1,
  ifAIC = TRUE,
  ifNCV = TRUE,
  kNCV = 0,
  NCV.nthreads = 1,
  pc = NULL,
  par.start,
  par.fix,
  upper.bound,
  CI = 0.95,
  CI.R = 1000,
  CI.seed = 123,
  eta = TRUE,
  tmp.file = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{the formula for the effect of exposure at time t regarding the
exposure history. Use sX() for the distributed lag term. For example y ~
sX(t, x).}

\item{smooth}{the formula for the other smooth functions. The smooths are
specified by mgcv::s. For example ~ s(z). Note also that the Gaussian
random effect models are accommodated, by setting s(bs = "re").}

\item{unpen.smooth}{the formula for the other unpenalized smooth functions.}

\item{fe.varying}{the formula for time-varying covariates in linear
components. For example ~ z}

\item{fe.cont}{the formula for time-invariant covariates in linear
components. For example ~ z}

\item{dat}{the data frame containing the variables.}

\item{maxL}{the maximum lag, default 14.}

\item{kE}{the dimension for B-spline for ACE response-exposure-function,
default 20.}

\item{kw}{the dimension for B-spline for weight function, default 20.}

\item{ifAIC}{whether or not to calculate the AIC, default TRUE.}

\item{pc}{the point constraint for ACE response-exposure-function, i.e. f(x0)
= 0. Use sum-to-zero constraint if \code{pc = NULL}. Default \code{pc =
NULL}.}

\item{par.start}{the starting values for BFGS in the outer optimization. If
the argument is missing, the starting values are the fitted values from
\code{mgcv::bam}.}

\item{par.fix}{the values of smoothing/dispersion parameters which is fixed
in the outer optimization. For example, set \code{par.fix = c(10,NA,NA)}
for fixing log(theta)=10 if the prior knowledge believes there is no
over-dispersion.}

\item{upper.bound}{the upper bound of parameters in BFGS.}

\item{CI}{the confidence level, default 0.95.}

\item{CI.R}{the number of sampling in the sampling method for CI, default
1000.}

\item{CI.seed}{the random seed in the sampling method for CI, default 123.}

\item{eta}{whether or not to report the CI for the linear predictor eta =
log(mu), default \code{TRUE}.}

\item{verbose}{whether or not to print the progress and diagnostic
information, such as the gradients and function values in inner and middle
optimizations.}
}
\value{
Object of class \code{drfDLNMadditive_fit}. S3 function \code{summary},
  \code{residuals}, and \code{predict}.
}
\description{
Title Fit a single index distributed response function distributed lag non-linear model
(SID-DLNM)
}
