% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_fit.R
\name{mgcv_fit}
\alias{mgcv_fit}
\title{Title fit a DRFDLNM using mgcv}
\usage{
mgcv_fit(formula.list, x.names.list, sXdat, kE, kw, verbose)
}
\arguments{
\item{verbose}{}
}
\description{
Title fit a DRFDLNM using mgcv
}
