% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCV.R
\name{NCV.drfDLNMadditive_fit}
\alias{NCV.drfDLNMadditive_fit}
\title{Title Obtain NCV from the fitted model by drfDLNMadditive}
\usage{
\method{NCV}{drfDLNMadditive_fit}(object, kNCV = 0, NCV.nthreads = 1, verbose = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{drfDLNMadditive_fit}.}

\item{kNCV}{}

\item{NCV.nthreads}{}

\item{verbose}{}

\item{...}{}
}
\description{
Title Obtain NCV from the fitted model by drfDLNMadditive
}
