% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateRatio.R
\name{RateRatio.drfDLNMadditive_fit}
\alias{RateRatio.drfDLNMadditive_fit}
\title{Title Rate Ratio for drfDLNMadditive_fit objects}
\usage{
\method{RateRatio}{drfDLNMadditive_fit}(object, x0, x1, verbose = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{drfDLNMadditive_fit}.}

\item{verbose}{whether to print messages during the process, default \code{FALSE}.}
}
\description{
Title Rate Ratio for drfDLNMadditive_fit objects
}
