% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.drfDLNMadditive_fit}
\alias{residuals.drfDLNMadditive_fit}
\title{Title Randomized Quantile Residuals}
\usage{
\method{residuals}{drfDLNMadditive_fit}(
  object,
  seed = 123,
  plot = FALSE,
  nsample = 10000,
  curve = FALSE,
  ...
)
}
\arguments{
\item{object}{object of class \code{drfDLNMadditive_fit}.}

\item{seed}{random seed, default 123.}

\item{plot}{whether or not to show the plots, default \code{FALSE}.}

\item{nsample}{the number of sampling from the datasets if number of
observations is larger than nsample, default \code{1e4}.}

\item{...}{}

\item{curver}{whether or not to draw a curve for residual plot using
\code{ggplot2::stat_smooth(method = "gam")}.}
}
\value{
a list containing the randomized quantile residuals and the plots.
}
\description{
Title Randomized Quantile Residuals
}
\references{
Dunn, Peter K., and Gordon K. Smyth. \dQuote{Randomized quantile
  residuals}. Journal of Computational and graphical statistics 5.3 (1996):
  236-244.
}
