% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.drfDLNMadditive_fit}
\alias{summary.drfDLNMadditive_fit}
\title{Title summary() Method for Objects of Class 'drfDLNMadditive_fit'}
\usage{
\method{summary}{drfDLNMadditive_fit}(
  object,
  E.eval.list,
  l.eval.list,
  others.eval = NULL,
  cen,
  plot = FALSE,
  true.function,
  ...
)
}
\arguments{
\item{object}{object of class \code{aceDLNM_fit}.}

\item{others.eval}{a data frame containing variables where the other smooth
functions are evaluated.}

\item{plot}{whether or not to show the plots, default \code{FALSE}.}

\item{true.function}{the list containing the true functions.
f(E) - f(E0).}

\item{...}{}

\item{E.eval}{a vector of ACE where the ACE response-exposure-function is
evaluated. The default is a sequence from 0.0025 quantile to 0.9975
quantile of estimated E, if the argument is missing.}
}
\description{
Title summary() Method for Objects of Class 'drfDLNMadditive_fit'
}
