% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{sX}
\alias{sX}
\title{Title formula for Effect of exposure process}
\usage{
sX(t, x, by)
}
\arguments{
\item{t}{time variable}

\item{x}{exposure variable}

\item{by}{}
}
\value{
a formula for model fitting in \code{drfDLNMadditive}.
}
\description{
Title formula for Effect of exposure process
}
